@@ -17,6 +17,8 @@ namespace Microsoft.Azure.IIoT.Services.OpcUa.History {
     using Microsoft.Azure.IIoT.Http.Default;
     using Microsoft.Azure.IIoT.Hub.Client;
     using Microsoft.Azure.IIoT.Module.Default;
+    using Microsoft.Azure.IIoT.Diagnostics;
+    using Microsoft.Extensions.Hosting;
     using Microsoft.Extensions.Configuration;
     using Microsoft.Extensions.DependencyInjection;
     using Microsoft.Extensions.Logging;
@@ -30,7 +32,6 @@ namespace Microsoft.Azure.IIoT.Services.OpcUa.History {
     using System;
     using Serilog;
     using ILogger = Serilog.ILogger;
-    using Microsoft.Azure.IIoT.Diagnostics;
 
     /// <summary>
     /// Webservice startup
@@ -50,7 +51,7 @@ namespace Microsoft.Azure.IIoT.Services.OpcUa.History {
         /// <summary>
         /// Current hosting environment - Initialized in constructor
         /// </summary>
-        public IHostingEnvironment Environment { get; }
+        public IWebHostEnvironment Environment { get; }
 
         /// <summary>
         /// Di container - Initialized in `ConfigureServices`
@@ -62,7 +63,7 @@ namespace Microsoft.Azure.IIoT.Services.OpcUa.History {
         /// </summary>
         /// <param name="env"></param>
         /// <param name="configuration"></param>
-        public Startup(IHostingEnvironment env, IConfiguration configuration) {
+        public Startup(IWebHostEnvironment env, IConfiguration configuration) {
             Environment = env;
             ServiceInfo = new ServiceInfo();
             Config = new Config(
@@ -104,10 +105,9 @@ namespace Microsoft.Azure.IIoT.Services.OpcUa.History {
             // services.AddHttpClient();
 
             // Add controllers as services so they'll be resolved.
-            services.AddMvc()
+            services.AddControllers()
                 .AddApplicationPart(GetType().Assembly)
-                .AddControllersAsServices()
-                .AddJsonOptions(options => {
+                .AddNewtonsoftJson(options => {
                     options.SerializerSettings.Formatting = Formatting.Indented;
                     options.SerializerSettings.Converters.Add(new ExceptionConverter(
                         Environment.IsDevelopment()));
@@ -119,6 +119,7 @@ namespace Microsoft.Azure.IIoT.Services.OpcUa.History {
                 Version = VersionInfo.PATH,
                 Description = ServiceInfo.Description,
             });
+            services.AddHealthChecks();
 
             // Prepare DI container
             var builder = new ContainerBuilder();
@@ -135,9 +136,12 @@ namespace Microsoft.Azure.IIoT.Services.OpcUa.History {
         /// </summary>
         /// <param name="app"></param>
         /// <param name="appLifetime"></param>
-        public void Configure(IApplicationBuilder app, IApplicationLifetime appLifetime) {
+        public void Configure(IApplicationBuilder app, IHostApplicationLifetime appLifetime) {
 
             var log = ApplicationContainer.Resolve<ILogger>();
+
+            app.UseRouting();
+            app.EnableCors();
 
             if (Config.AuthRequired) {
                 app.UseAuthentication();
@@ -147,15 +151,16 @@ namespace Microsoft.Azure.IIoT.Services.OpcUa.History {
                 app.UseHttpsRedirection();
             }
 
-            app.EnableCors();
-
             app.UseSwagger(Config, new Info {
                 Title = ServiceInfo.Name,
                 Version = VersionInfo.PATH,
                 Description = ServiceInfo.Description,
             });
 
-            app.UseMvc();
+            app.UseEndpoints(endpoints => {
+                endpoints.MapHealthChecks("/health");
+                endpoints.MapControllers();
+            });
 
             // If you want to dispose of resources that have been resolved in the
             // application container, register for the "ApplicationStopped" event.
